Move procrastination detector to service package
