+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:theme="@style/Theme.Multimatum">
 
         <service
-            android:name=".procrastination_detector.ProcrastinationDetectorService"
+            android:name=".service.ProcrastinationDetectorService"
             android:enabled="true"
             android:exported="false" />
         <activity
